+++ packages/blocks/core/js/wordpress/home-link/index.js
@@ -1,9 +1,17 @@
 // @flow
+/**
+ * External dependencies
+ */
+import { __ } from '@wordpress/i18n';
 
 /**
  * Blockera dependencies
  */
-import { SharedBlockExtension } from '@blockera/editor';
+import {
+	SharedBlockExtension,
+	generalBlockStates,
+	sharedBlockStates,
+} from '@blockera/editor';
 
 /**
  * Internal dependencies
@@ -15,7 +23,46 @@ export const HomeLink: BlockType = {
 	name: 'blockeraHomeLink',
 	targetBlock: 'core/home-link',
 	blockeraInnerBlocks: {
-		'elements/link': sharedInnerBlocks['elements/link'],
+		'elements/link': {
+			...sharedInnerBlocks['elements/link'],
+			label: __('Link', 'blockera'),
+			description: __('Hyperlink element.', 'blockera'),
+			settings: {
+				...sharedInnerBlocks['elements/link'].settings,
+				force: false,
+			},
+		},
+	},
+	availableBlockStates: {
+		...generalBlockStates,
+		'current-menu-item': {
+			type: 'current-menu-item',
+			label: __('Current Page', 'blockera'),
+			category: 'special',
+			tooltip: (
+				<>
+					<h5>{__('Is Current Page?', 'blockera')}</h5>
+					<p>
+						{__(
+							'Activates if the block URL is the same as the current page URL.',
+							'blockera'
+						)}
+					</p>
+					<code style={{ margin: '5px 0' }}>
+						<span style={{ opacity: '0.7' }}>.block</span>
+						.current-menu-item
+					</code>
+				</>
+			),
+			breakpoints: {},
+			priority: 10,
+			force: true,
+			settings: {
+				color: 'var(--blockera-controls-states-color)',
+			},
+		},
+		active: sharedBlockStates.active,
+		visited: sharedBlockStates.visited,
 	},
 	edit: (props) => {
 		return <SharedBlockExtension {...props} />;